requirements.txt
git+https://github.com/Rapptz/discord.py
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
chardet==4.0.0
colorama==0.4.4
decorator==5.0.9
discord.py==1.7.2
idna==3.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
multidict==5.1.0
parso==0.8.2
pickleshare==0.7.5
profanity==1.1
prompt-toolkit==3.0.18
Pygments==2.9.0
python-dateutil==2.8.1
pyzmq==22.0.3
six==1.16.0
tk==0.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
yarl==1.6.3
